next build succeeds
Changes:
- refactored nextauth
- added postcss-nesting
- limited font imports
